requirements.txt
langchain
langchain_text_splitters 
langchain-google-genai
langchain-community 
pypdf
streamlit
unstructured
langchain-ollama
unstructured[all]
# WebBaseLoader
requests
beautifulsoup4
lxml
python-dotenv
pypdf